+++ vpin-studio-ui/src/main/java/de/mephisto/vpin/ui/tables/TableOverviewContextMenu.java
@@ -78,7 +78,7 @@ public class TableOverviewContextMenu {
       ctxMenu.getItems().add(vpsItem);
 
       MenuItem vpsUpdateItem = new MenuItem("Reset VPS Updates");
-      vpsUpdateItem.setOnAction(actionEvent -> tableOverviewController.onVps());
+      vpsUpdateItem.setOnAction(actionEvent -> tableOverviewController.onVpsReset());
       vpsUpdateItem.setDisable(game.getVpsUpdates().isEmpty());
       vpsUpdateItem.setGraphic(iconVpsReset);
       ctxMenu.getItems().add(vpsUpdateItem);